requirements.txt
absl-py
astor
bleach
bottle
click
certifi
chardet
flask
future
gast
grpcio
gunicorn
h5py
html5lib
itsdangerous
idna
Jinja2
Keras>=3.0.0
line-bot-sdk
Markdown
MarkupSafe
numpy<2.0.0
oauthlib
pillow
protobuf
PyYAML
python-dotenv
requests
scipy
six
tensorboard
tensorflow-cpu
termcolor
urllib3
Werkzeug